Get bundler options for BundlerManager#update_gem
This way the method is generic to work with any options we may want. Or
none.

The goal is to achieve a update which first runs in conservative mode
and later with the default update. Like this the git commits are as
small as possible, but we also keep a high rate of automatic updates.

An future feature here is to receive through the configuration file any
options the user may want.