Fix changes being listed in `git status` but that were already commited
This was causing confusion among users (me included) about wether
gisdatigo has commited everything it was supposed to or that the diff
left must be added to a manual commit.

Creating commits using libgit involves a lot more than what we are used
to deal with when using `git add . && git commit`. To honest I still
don't understand what was happening to make changes already commited to
still be listed as part of the `git status` output.

The tip for the the fix came from https://stackoverflow.com/a/24558847.