Add executable CLI
Before this, the only supported way of running gisdatigo was by using a
rake task loaded as a railtie. This made things a bit less
straightforward for running, since we had to add the gisdatigo gem to
the Gemfile to be able to run, and in case we wanted to use it ti update
a non rails application, there was no way to do it. Now, even in a non
rails app, running 'gem install gisdatigo;gisdatigo' should be enough to
get things running.

Part of #10.